Timestamp fix, driver divided into 3 separate drivers
