+++ src/test/it/resources/basic_analytics/stb_general/ethernet_stats_reports/expected_result.txt
@@ -1,0 +1,3 @@
+{"@timestamp": "2018-04-24T09:20:00.000Z", "metric_name": "stb_ethernet_report.ethernet_network.viewer_id.distinct_count", "value": 2}
+{"@timestamp": "2018-04-24T09:20:00.000Z", "metric_name": "stb_ethernet_report.ethernet.average_downstream_kbps.rxKbps.avg", "value": 33.33}
+{"@timestamp": "2018-04-24T09:20:00.000Z", "metric_name": "stb_ethernet_report.ethernet.average_upstream_kbps.txKbps.avg", "value": 33.33}
