+++ src/test/it/resources/basic_analytics/stb_general/ethernet_stats_reports/input/ethernet_reports.txt
@@ -1,0 +1,11 @@
+{"@timestamp":"2018-04-24T09:27:46.248Z","EthernetStats":{"iface":"eth2.link","rxPackets":1557049,"rxKbps":100,"txPackets":430163,"type":"EthernetStats","txKbps":0,"ts":1524562010500},"ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524562065934},"type":"eosstb_log_parsing_flat"},"fields":{"@timestamp":[1524562066248]}
+{"@timestamp":"2018-04-24T09:27:46.248Z","EthernetStats":{"iface":"eth2.link","rxPackets":1557049,"rxKbps":0,"txPackets":430163,"type":"EthernetStats","txKbps":100,"ts":1524562010500},"ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295302","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524562065934},"type":"eosstb_log_parsing_flat"},"fields":{"@timestamp":[1524562066248]}
+{"@timestamp":"2018-04-24T09:27:46.248Z","EthernetStats":{"iface":"eth2.link","rxPackets":1557049,"rxKbps":0,"txPackets":430163,"type":"EthernetStats","txKbps":0,"ts":1524562010500},"ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295303","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524562065934},"type":"eosstb_log_parsing_flat"},"fields":{"@timestamp":[1524562066248]}
+{"@timestamp":"2018-05-24T09:27:46.248Z","EthernetStats":{"iface":"eth2.link","rxPackets":1557049,"rxKbps":100,"txPackets":430163,"type":"EthernetStats","txKbps":100,"ts":1524562010500},"ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524562065934},"type":"eosstb_log_parsing_flat"},"fields":{"@timestamp":[1524562066248]}
+{"@timestamp":"2018-05-24T09:27:46.248Z","EthernetStats":{"iface":"eth2.link","rxPackets":1557049,"rxKbps":0,"txPackets":430163,"type":"EthernetStats","txKbps":0,"ts":1524562010500},"ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524562065934},"type":"eosstb_log_parsing_flat"},"fields":{"@timestamp":[1524562066248]}
+{"@timestamp":"2018-05-24T09:27:46.248Z","EthernetStats":{"iface":"eth2.link","rxPackets":1557049,"rxKbps":0,"txPackets":430163,"type":"EthernetStats","txKbps":0,"ts":1524562010500},"ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295302","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524562065934},"type":"eosstb_log_parsing_flat"},"fields":{"@timestamp":[1524562066248]}
+
+
+
+
+
