--- src/test/it/resources/basic_analytics/stb_general/ethernet_wifi_reports/expected_result.txt
@@ -1,10 +1,0 @@
-{"@timestamp": "2018-04-24T08:30:00.000Z", "metric_name": "stb_ethernet_wifi_report.wireless.average_downstream_kbps.rxKbps.avg", "value": 33.33}
-{"@timestamp": "2018-04-24T09:20:00.000Z", "metric_name": "stb_ethernet_wifi_report.ethernet.average_upstream_kbps.txKbps.avg", "value": 33.33}
-{"@timestamp": "2018-04-24T09:20:00.000Z", "metric_name": "stb_ethernet_wifi_report.ethernet_network.viewer_id.distinct_count", "value": 2}
-{"@timestamp": "2018-04-24T11:00:00.000Z", "metric_name": "stb_ethernet_wifi_report.total_cpe_net_config.net_configuration_type.Ethernet.viewer_id.distinct_count", "value": 2, "net_configuration_type": "Ethernet"}
-{"@timestamp": "2018-04-24T11:00:00.000Z", "metric_name": "stb_ethernet_wifi_report.total_cpe_net_config.net_configuration_type.WiFi.viewer_id.distinct_count", "value": 3, "net_configuration_type": "WiFi"}
-{"@timestamp": "2018-04-24T11:00:00.000Z", "value": 2, "net_config_enabled": "1", "metric_name": "stb_ethernet_wifi_report.stb_with.net_config_enabled.1.viewer_id.distinct_count"}
-{"@timestamp": "2018-04-24T11:00:00.000Z", "value": 3, "net_config_enabled": "0", "metric_name": "stb_ethernet_wifi_report.stb_with.net_config_enabled.0.viewer_id.distinct_count"}
-{"@timestamp": "2018-04-24T08:30:00.000Z", "metric_name": "stb_ethernet_wifi_report.wifi_network.viewer_id.distinct_count", "value": 2}
-{"@timestamp": "2018-04-24T08:30:00.000Z", "metric_name": "stb_ethernet_wifi_report.wireless.average_upstream_kbps.txKbps.avg", "value": 33.33}
-{"@timestamp": "2018-04-24T09:20:00.000Z", "metric_name": "stb_ethernet_wifi_report.ethernet.average_downstream_kbps.rxKbps.avg", "value": 33.33}
