+++ src/test/it/resources/basic_analytics/stb_general/net_configuration_reports/configuration.yml
@@ -1,0 +1,11 @@
+spark:
+  appName: stb_ethernet_wifi_report-basic-analytics-test
+  master: local[*]
+  shuffle.partitions: 1
+  outputMode: update
+#  consoleWriter: true
+
+analytics:
+  componentName: stb_net_config_report
+  watermark: '1 minutes'
+  window: '10 minutes'
