+++ src/test/it/resources/basic_analytics/stb_general/net_configuration_reports/input/net_configuration.txt
@@ -1,0 +1,10 @@
+{"@timestamp":"2018-04-24T11:03:19.350Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524567799049},"type":"eosstb_log_parsing_flat","NetConfiguration":{"ifaces":[{"iface":"eth2","ipv4":{"router":"192.168.0.1","ip":"192.168.0.242","type":"IPNetwork","ts":1524567707159,"mask":"255.255.255.0"},"ipv6":{"router":"fe80::925c:44ff:fe6e:397e","ip":"2a02:1810:1c23:6500:929d:7dff:fe2f:bacd","type":"IPNetwork","ts":1524567707159,"mask":"64"},"type":"Ethernet","enabled":1}],"type":"NetConfiguration","ts":1524567707159}}
+{"@timestamp":"2018-04-24T11:03:19.350Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295302","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524567799049},"type":"eosstb_log_parsing_flat","NetConfiguration":{"ifaces":[{"iface":"eth2","ipv4":{"router":"192.168.0.1","ip":"192.168.0.242","type":"IPNetwork","ts":1524567707159,"mask":"255.255.255.0"},"ipv6":{"router":"fe80::925c:44ff:fe6e:397e","ip":"2a02:1810:1c23:6500:929d:7dff:fe2f:bacd","type":"IPNetwork","ts":1524567707159,"mask":"64"},"type":"Ethernet","enabled":1}],"type":"NetConfiguration","ts":1524567707159}}
+{"@timestamp":"2018-04-24T11:03:19.350Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295303","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524567799049},"type":"eosstb_log_parsing_flat","NetConfiguration":{"ifaces":[{"iface":"eth2","ipv4":{"router":"192.168.0.1","ip":"192.168.0.242","type":"IPNetwork","ts":1524567707159,"mask":"255.255.255.0"},"ipv6":{"router":"fe80::925c:44ff:fe6e:397e","ip":"2a02:1810:1c23:6500:929d:7dff:fe2f:bacd","type":"IPNetwork","ts":1524567707159,"mask":"64"},"type":"WiFi","enabled":0}],"type":"NetConfiguration","ts":1524567707159}}
+{"@timestamp":"2018-05-24T11:03:19.350Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524567799049},"type":"eosstb_log_parsing_flat","NetConfiguration":{"ifaces":[{"iface":"eth2","ipv4":{"router":"192.168.0.1","ip":"192.168.0.242","type":"IPNetwork","ts":1524567707159,"mask":"255.255.255.0"},"ipv6":{"router":"fe80::925c:44ff:fe6e:397e","ip":"2a02:1810:1c23:6500:929d:7dff:fe2f:bacd","type":"IPNetwork","ts":1524567707159,"mask":"64"},"type":"Ethernet","enabled":1}],"type":"NetConfiguration","ts":1524567707159}}
+{"@timestamp":"2018-05-24T11:03:19.350Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524567799049},"type":"eosstb_log_parsing_flat","NetConfiguration":{"ifaces":[{"iface":"eth2","ipv4":{"router":"192.168.0.1","ip":"192.168.0.242","type":"IPNetwork","ts":1524567707159,"mask":"255.255.255.0"},"ipv6":{"router":"fe80::925c:44ff:fe6e:397e","ip":"2a02:1810:1c23:6500:929d:7dff:fe2f:bacd","type":"IPNetwork","ts":1524567707159,"mask":"64"},"type":"WiFi","enabled":0}],"type":"NetConfiguration","ts":1524567707159}}
+{"@timestamp":"2018-05-24T11:03:19.350Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295302","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524567799049},"type":"eosstb_log_parsing_flat","NetConfiguration":{"ifaces":[{"iface":"eth2","ipv4":{"router":"192.168.0.1","ip":"192.168.0.242","type":"IPNetwork","ts":1524567707159,"mask":"255.255.255.0"},"ipv6":{"router":"fe80::925c:44ff:fe6e:397e","ip":"2a02:1810:1c23:6500:929d:7dff:fe2f:bacd","type":"IPNetwork","ts":1524567707159,"mask":"64"},"type":"WiFi","enabled":0}],"type":"NetConfiguration","ts":1524567707159}}
+
+
+
+
