+++ src/test/it/resources/basic_analytics/stb_general/wifi_stats_reports/configuration.yml
@@ -6,6 +6,6 @@ spark:
 #  consoleWriter: true
 
 analytics:
-  componentName: stb_ethernet_wifi_report
+  componentName: st_wifi_report
   watermark: '1 minutes'
   window: '10 minutes'
