+++ src/test/it/resources/basic_analytics/stb_general/wifi_stats_reports/input/wifi_reports.txt
@@ -1,0 +1,6 @@
+{"@timestamp":"2018-04-24T08:43:19.222Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524559398723},"WiFiStats":{"iface":"wifi.radio.1","rxPackets":0,"RSSi":[0,0,0],"rxKbps":100,"txPackets":0,"type":"WiFiStats","txKbps":0,"ts":1524559193512},"type":"eosstb_log_parsing_flat"}
+{"@timestamp":"2018-04-24T08:43:19.222Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295302","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524559398723},"WiFiStats":{"iface":"wifi.radio.1","rxPackets":0,"RSSi":[0,0,0],"rxKbps":0,"txPackets":0,"type":"WiFiStats","txKbps":100,"ts":1524559193512},"type":"eosstb_log_parsing_flat"}
+{"@timestamp":"2018-04-24T08:43:19.222Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295303","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524559398723},"WiFiStats":{"iface":"wifi.radio.1","rxPackets":0,"RSSi":[0,0,0],"rxKbps":0,"txPackets":0,"type":"WiFiStats","txKbps":0,"ts":1524559193512},"type":"eosstb_log_parsing_flat"}
+{"@timestamp":"2018-05-24T08:43:19.222Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524559398723},"WiFiStats":{"iface":"wifi.radio.1","rxPackets":0,"RSSi":[0,0,0],"rxKbps":100,"txPackets":0,"type":"WiFiStats","txKbps":100,"ts":1524559193512},"type":"eosstb_log_parsing_flat"}
+{"@timestamp":"2018-05-24T08:43:19.222Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295301","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524559398723},"WiFiStats":{"iface":"wifi.radio.1","rxPackets":0,"RSSi":[0,0,0],"rxKbps":0,"txPackets":0,"type":"WiFiStats","txKbps":0,"ts":1524559193512},"type":"eosstb_log_parsing_flat"}
+{"@timestamp":"2018-05-24T08:43:19.222Z","ODH_NTPSynced":"true","@version":"1","header":{"anonymizedData":false,"profileName":"","modelName":"DCX960","privacyLevel":"PUBLIC","viewerID":"3C36E4-EOSSTB-003479295302","hardwareVersion":"ARRIS-DCX960-MPA+","softwareVersions":[{"component":"app","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"as","version":"4.8_16_20180328102331_master_eos_sprint39"},{"component":"firmware","version":"DCX960__-mon-prd-00.01-039-ah-AL-20180328102331-na004"}],"ts":1524559398723},"WiFiStats":{"iface":"wifi.radio.1","rxPackets":0,"RSSi":[0,0,0],"rxKbps":0,"txPackets":0,"type":"WiFiStats","txKbps":0,"ts":1524559193512},"type":"eosstb_log_parsing_flat"}
